refactor option.h and related
