+++ src/core/CMakeLists.txt
@@ -19,9 +19,12 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/matcher-factory.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/motion.h"
         "${CMAKE_CURRENT_LIST_DIR}/motion-frame.h"
+        "${CMAKE_CURRENT_LIST_DIR}/notification.h"
         "${CMAKE_CURRENT_LIST_DIR}/video.h"
         "${CMAKE_CURRENT_LIST_DIR}/video-frame.h"
-        "${CMAKE_CURRENT_LIST_DIR}/options.h"
+        "${CMAKE_CURRENT_LIST_DIR}/options-container.h"
+        "${CMAKE_CURRENT_LIST_DIR}/options-container.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/options-interface.h"
         "${CMAKE_CURRENT_LIST_DIR}/options-registry.h"
         "${CMAKE_CURRENT_LIST_DIR}/options-registry.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/info.h"
