+++ src/core/processing-block-interface.h
@@ -3,7 +3,6 @@
 
 #pragma once
 
-#include "options.h"
 #include "info-interface.h"
 #include <src/types.h>
