+++ src/ds/d500/d500-options.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "ds/ds-private.h"
-#include "core/options.h"
+#include "core/options-container.h"
 #include "option.h"
 
 #include <rsutils/lazy.h>
