+++ src/error-handling.h
@@ -2,11 +2,16 @@
 /* Copyright(c) 2019 Intel Corporation. All Rights Reserved. */
 #pragma once
 
-#include "option.h"
-#include "types.h"
+#include "core/option-interface.h"
+#include <rsutils/concurrency/concurrency.h>
+
 
 namespace librealsense
 {
+    class notification_decoder;
+    class notifications_processor;
+
+
     class polling_error_handler
     {
     public:
@@ -30,6 +35,41 @@ namespace librealsense
         std::shared_ptr < active_object<> > _active_object;
         std::weak_ptr<notifications_processor> _notifications_processor;
         std::shared_ptr<notification_decoder> _decoder;
+    };
+
+
+    class polling_errors_disable : public option
+    {
+    public:
+        polling_errors_disable( std::shared_ptr< polling_error_handler > handler )
+            : _polling_error_handler( handler )
+            , _value( 1 )
+        {
+        }
+
+        ~polling_errors_disable();
+
+        void set( float value ) override;
+
+        float query() const override;
+
+        option_range get_range() const override;
+
+        bool is_enabled() const override;
+
+
+        const char * get_description() const override;
+
+        const char * get_value_description( float value ) const override;
+        void enable_recording( std::function< void( const option & ) > record_action ) override
+        {
+            _recording_function = record_action;
+        }
+
+    private:
+        std::weak_ptr< polling_error_handler > _polling_error_handler;
+        float _value;
+        std::function< void( const option & ) > _recording_function = []( const option & ) {};
     };
 
 }
