+++ src/linux/backend-v4l2.cpp
@@ -4,6 +4,7 @@
 #include "backend-v4l2.h"
 #include <src/platform/command-transfer.h>
 #include <src/core/time-service.h>
+#include <src/core/notification.h>
 #include "backend-hid.h"
 #include "backend.h"
 #include "types.h"
