+++ wrappers/python/CMakeLists.txt
@@ -46,6 +46,7 @@ set(RAW_RS
     ../../src/frame.cpp
     ../../src/environment.cpp
     ../../src/to-string.cpp
+    ../../src/core/options-container.cpp
     ../../src/core/options-registry.cpp
 )
