+++ wrappers/python/pybackend.cpp
@@ -482,4 +482,3 @@ void librealsense::info_container::enable_recording(std::function<void(const inf
 void librealsense::info_container::update(std::shared_ptr<extension_snapshot> ext){}
 bool librealsense::info_container::supports_info(rs2_camera_info info) const { return false; }
 const std::string& librealsense::info_container::get_info(enum rs2_camera_info) const { static std::string s = ""; return s; }
-std::vector<rs2_option> librealsense::options_container::get_supported_options(void)const { return{}; }
